LU-4629 lnet: fix issues found by Klocwork Insight tool
Null pointer 'cp' that comes from line 2544 may be dereferenced
at line 2618.

Pointer 'ni' checked for NULL at line 1569 may be passed to
function and may be dereferenced there by passing argument 1 to
function 'lnet_ni_notify_locked' at line 1621.

Null pointer 'best_iface' that comes from line 802 may be
dereferenced at line 832.

Buffer overflow of string buffer due to non null terminated string.

sscanf format specification '%lx' expects type 'unsigned long int *'
for 'x', but parameter 4 has a different type 'long int*'.

Pointer 'tsc' returned from call to function 'sfw_find_test_case'
at line 571 may be NULL and will be dereferenced at line 572.

sscanf format specification '%i' expects type 'int *' for 'i',
but parameter 3 has a different type 'unsigned int*'.

Local variable 'hash' is never used.

+++ lnet/utils/portals.c
@@ -1049,7 +1049,7 @@ jt_ptl_fail_nid (int argc, char **argv)
 {
         int                      rc;
         lnet_nid_t               nid;
-        unsigned int             threshold;
+	int			 threshold;
         struct libcfs_ioctl_data data;
 
         if (argc < 2 || argc > 3)
@@ -1066,14 +1066,14 @@ jt_ptl_fail_nid (int argc, char **argv)
 
         if (argc < 3) {
                 threshold = LNET_MD_THRESH_INF;
-        } else if (sscanf (argv[2], "%i", &threshold) != 1) {
+	} else if (sscanf(argv[2], "%i", &threshold) != 1) {
                 fprintf (stderr, "Can't parse count \"%s\"\n", argv[2]);
                 return (-1);
         }
 
         LIBCFS_IOC_INIT (data);
         data.ioc_nid = nid;
-        data.ioc_count = threshold;
+	data.ioc_count = (__u32)threshold;
 
         rc = l_ioctl (LNET_DEV_ID, IOC_LIBCFS_FAIL_NID, &data);
         if (rc < 0)
